refactor: separar os estilos da lógica do componente
